Policies de artigos

+++ app/Http/Controllers/ArticleController.php
@@ -10,6 +10,7 @@ use App\Services\ResponseService;
 use Symfony\Component\HttpFoundation\Response;
 use App\Http\Requests\ArticleStoreRequest;
 use App\Http\Requests\ArticleUpdateRequest;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller
 {
@@ -18,12 +19,14 @@ class ArticleController extends Controller
 
     public function index()
     {
+        Gate::authorize('index', Article::class);
         $articles = Article::all();
         return ResponseService::success('Artigos encontrados com sucesso', Response::HTTP_OK, $articles);
     }
 
     public function indexDraft()
     {
+        Gate::authorize('indexDraft', Article::class);
         $articles = Article::where('status_id', StatusEnum::DRAFT)->get();
         return ResponseService::success('Artigos em rascunho encontrados com sucesso', Response::HTTP_OK, $articles);
     }
@@ -50,6 +53,7 @@ class ArticleController extends Controller
 
     public function inReview(Article $article)
     {
+        Gate::authorize('inReview', Article::class);
         $article->status_id = StatusEnum::IN_REVIEW;
         $article->save();
         return ResponseService::success('Artigo enviado para revisão com sucesso', Response::HTTP_OK);
@@ -57,6 +61,7 @@ class ArticleController extends Controller
 
     public function reviewed(Article $article)
     {
+        Gate::authorize('reviewed', Article::class);
         $article->status_id = StatusEnum::REVIEWED;
         $article->save();
         return ResponseService::success('Artigo revisado com sucesso', Response::HTTP_OK);
@@ -64,6 +69,7 @@ class ArticleController extends Controller
 
     public function published(Article $article)
     {
+        Gate::authorize('published', Article::class);
         $article->status_id = StatusEnum::PUBLISHED;
         $article->save();
         return ResponseService::success('Artigo publicado com sucesso', Response::HTTP_OK);
@@ -77,6 +83,7 @@ class ArticleController extends Controller
 
     public function store(ArticleStoreRequest $request)
     {
+        Gate::authorize('store', Article::class);
         Article::create($request->all());
         return ResponseService::success('Artigo criado com sucesso', Response::HTTP_CREATED);
     }
@@ -89,6 +96,7 @@ class ArticleController extends Controller
 
     public function destroy(Article $article)
     {
+        Gate::authorize('destroy', Article::class);
         $article->delete();
         return ResponseService::success('Artigo deletado com sucesso', Response::HTTP_OK);
     }
